+++ CHANGES.rst
@@ -1,3 +1,9 @@
+0.1.6 (2016-03-05)
+------------------
+Add HostsManager Mixin
+Create Lockfile utility
+move Timeout utility in utils.py
+
 0.1.5 (2016-03-23)
 ------------------
 Add Gauges Mixin